+++ CHANGELOG.md
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- Added additional modern web parts to our enums (CallToAction, Button, Sites)
+
+### Changed
+
 ## [3.13.1909.0 - September 2019 release]
 
 ### Added